+++ backend/src/database_functions/test_frontend.ts
@@ -1,0 +1,8 @@
+import { Request, Response } from "express";
+
+const handleFrontendSignal = (req: Request, res: Response) => {
+    console.log('signal received.');
+    res.status(201).json({status: 200, data: "Success!!!"});
+}
+
+module.exports = handleFrontendSignal;