+++ frontend/src/app.ts
@@ -1,5 +1,7 @@
 import P5 from "p5";
 
+import { signalServer } from "./server_functions";
+
 import "./styles.scss";
 
 const sketch = (p5:P5) => {
@@ -8,6 +10,11 @@ const sketch = (p5:P5) => {
         const canvas = p5.createCanvas(512, 1024);
         canvas.parent("app");
         p5.background("black");
+    }
+
+    p5.mouseClicked = () => {
+        // On any click, send signal to backend:
+        signalServer();
     }
 
     p5.draw = () => {